+++ branches/mod/equicord/main.js
@@ -1,0 +1,1 @@
+require("./equicord-desktop/equibopMain.js");