+++ branches/mod/shelter/selector-ui.js
@@ -47,7 +47,8 @@
 
 	const updateCurrent = () => SheltupdateNative.getCurrentBranches().then(setCurrentBranches);
 	updateCurrent().then(() => {
-		if (window.Vencord && !currentBranches().includes("vencord")) setVencordOtherwiseLoaded(true);
+		if (window.Vencord && !currentBranches().includes("vencord") && !currentBranches().includes("equicord"))
+			setVencordOtherwiseLoaded(true);
 
 		if (window.BdApi && !currentBranches().includes("betterdiscord")) setBdOtherwiseLoaded(true);
 	});
@@ -91,8 +92,8 @@
 			if (props.name === "shelter" && props.value) {
 				return "You need shelter to have access to this menu. Try uninstalling sheltupdate.";
 			}
-			if (props.name === "vencord" && vencordOtherwiseLoaded()) {
-				return "Vencord is currently loaded by some other mechanism.";
+			if ((props.name === "vencord" || props.name === "equicord") && vencordOtherwiseLoaded()) {
+				return "Vencord or Equicord are currently loaded by some other mechanism.";
 			}
 			if (props.name === "betterdiscord" && bdOtherwiseLoaded()) {
 				return "BetterDiscord is currently loaded by some other mechanism.";